refactor!: makefile and project structure
