+++ .gitignore
@@ -3,7 +3,7 @@ control
 machine-emulator-tools-*.tar.gz
 machine-emulator-tools-*.deb
 rootfs.*
-libcmt*
+libcmt-*.*
 
 # Prerequisites
 *.d
