+++ sys-utils/libcmt/.gitignore
@@ -1,0 +1,1 @@
+compile_flags.txt
