+++ sys-utils/libcmt/src/abi.c
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "abi.h"
+
 #include <errno.h>
-#include <libcmt/abi.h>
 #include <string.h>
 
 static uintptr_t align_forward(uintptr_t p, size_t a) {
