+++ sys-utils/libcmt/src/abi.h
