+++ sys-utils/libcmt/src/buf.h
