+++ sys-utils/libcmt/src/libcmt_mock.pc
