+++ sys-utils/libcmt/src/merkle.h
