+++ sys-utils/libcmt/tests/abi-single.c
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "libcmt/abi.h"
+#include "abi.h"
+
 #include <assert.h>
 #include <errno.h>
 #include <string.h>
