+++ sys-utils/libcmt/tests/keccak.c
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "abi.h"
+#include "keccak.h"
+
 #include <assert.h>
 #include <errno.h>
-#include <libcmt/abi.h>
-#include <libcmt/keccak.h>
 #include <stdio.h>
 #include <string.h>
