+++ sys-utils/libcmt/tests/merkle.c
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "merkle.h"
+
 #include <assert.h>
-#include <libcmt/merkle.h>
 #include <stdio.h>
 #include <string.h>
