+++ sys-utils/libcmt/tests/rollup_echo.c
@@ -1,3 +1,25 @@
+/* Copyright Cartesi and individual authors (see AUTHORS)
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "rollup.h"
+
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
 /* Data encoding:
  * - cast calldata "EvmAdvance(address,uint256,uint256,uint256,bytes)" \
  *   0x0000000000000000000000000000000000000000 \
@@ -10,12 +32,6 @@
  * - cast calldata-decode "Voucher(address,uint256,bytes)" 0x`xxd -p -c0 "<output.bin>"`
  *
  */
-#include "libcmt/rollup.h"
-#include <errno.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
 int main(void) {
     cmt_rollup_t rollup;
